feat: 이미지 등록 Api 및 컨트롤러 구현

+++ src/main/java/com/ridetogether/server/domain/image/application/OracleImageService.java
@@ -1,7 +1,12 @@
 package com.ridetogether.server.domain.image.application;
 
+import com.oracle.bmc.ConfigFileReader;
+import com.oracle.bmc.ConfigFileReader.ConfigFile;
 import com.oracle.bmc.Region;
+import com.oracle.bmc.auth.AuthenticationDetailsProvider;
+import com.oracle.bmc.auth.ConfigFileAuthenticationDetailsProvider;
 import com.oracle.bmc.objectstorage.ObjectStorage;
+import com.oracle.bmc.objectstorage.ObjectStorageClient;
 import com.oracle.bmc.objectstorage.model.CreatePreauthenticatedRequestDetails;
 import com.oracle.bmc.objectstorage.model.CreatePreauthenticatedRequestDetails.AccessType;
 import com.oracle.bmc.objectstorage.requests.CreatePreauthenticatedRequestRequest;
@@ -11,20 +16,23 @@ import com.oracle.bmc.objectstorage.requests.GetPreauthenticatedRequestRequest;
 import com.oracle.bmc.objectstorage.requests.PutObjectRequest;
 import com.oracle.bmc.objectstorage.responses.CreatePreauthenticatedRequestResponse;
 import com.oracle.bmc.objectstorage.responses.GetPreauthenticatedRequestResponse;
+import com.oracle.bmc.objectstorage.transfer.UploadConfiguration;
 import com.oracle.bmc.objectstorage.transfer.UploadManager;
 import com.oracle.bmc.objectstorage.transfer.UploadManager.UploadRequest;
 import com.oracle.bmc.objectstorage.transfer.UploadManager.UploadResponse;
 import com.ridetogether.server.domain.image.dao.ImageRepository;
 import com.ridetogether.server.domain.image.domain.Image;
+import com.ridetogether.server.domain.image.model.ImageType;
 import com.ridetogether.server.domain.member.dao.MemberRepository;
 import com.ridetogether.server.domain.member.domain.Member;
 import com.ridetogether.server.global.apiPayload.code.status.ErrorStatus;
-import com.ridetogether.server.global.apiPayload.exception.handler.MemberHandler;
+import com.ridetogether.server.global.apiPayload.exception.handler.ErrorHandler;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -32,7 +40,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -42,10 +49,6 @@ import org.springframework.web.multipart.MultipartFile;
 @Slf4j
 @Transactional
 public class OracleImageService implements ImageService {
-
-	private final ObjectStorage client;
-
-	private final UploadManager uploadManager;
 
 	private final MemberRepository memberRepository;
 
@@ -62,12 +65,38 @@ public class OracleImageService implements ImageService {
 
 	private static final long PRE_AUTH_EXPIRE_MINUTE = 20;
 
+	public ObjectStorage getClient() throws Exception {
+		ConfigFile config = ConfigFileReader.parse("~/.oci/config", "DEFAULT");
+
+		AuthenticationDetailsProvider provider = new ConfigFileAuthenticationDetailsProvider(config);
+
+		return ObjectStorageClient.builder()
+				.region(Region.AP_CHUNCHEON_1)
+				.build(provider);
+	}
+
+	public UploadManager getManager(ObjectStorage client) throws Exception {
+		UploadConfiguration configuration = UploadConfiguration.builder()
+				.allowMultipartUploads(true)
+				.allowParallelUploads(true)
+				.build();
+		return new UploadManager(client, configuration);
+	}
+
+	public UploadConfiguration getUploadConfiguration() {
+		//upload object
+		return UploadConfiguration.builder()
+				.allowMultipartUploads(true)
+				.allowParallelUploads(true)
+				.build();
+	}
+
 	@Override
-	public Long uploadProfileImg(MultipartFile file, Long memberIdx) throws Exception{
+	public Long uploadProfileImg(MultipartFile file, Long memberIdx) throws Exception {
 		File uploadFile = convert(file)  // 파일 변환할 수 없으면 에러
 				.orElseThrow(() -> new IllegalArgumentException("error: MultipartFile -> File convert fail"));
 		String fileDir = memberIdx + "/" + PROFILE_IMG_DIR;
-		Member member = memberRepository.findByIdx(memberIdx).orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+		Member member = memberRepository.findByIdx(memberIdx).orElseThrow(() -> new ErrorHandler(ErrorStatus.MEMBER_NOT_FOUND));
 		return upload(uploadFile, fileDir, member);
 	}
 	@Override
@@ -75,13 +104,19 @@ public class OracleImageService implements ImageService {
 		File uploadFile = convert(file)  // 파일 변환할 수 없으면 에러
 				.orElseThrow(() -> new IllegalArgumentException("error: MultipartFile -> File convert fail"));
 		String fileDir = memberIdx + "/" + KAKAO_IMG_DIR;
-		Member member = memberRepository.findByIdx(memberIdx).orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+		Member member = memberRepository.findByIdx(memberIdx).orElseThrow(() -> new ErrorHandler(ErrorStatus.MEMBER_NOT_FOUND));
 		return upload(uploadFile, fileDir, member);
 	}
 
 	@Override
-	public String getPublicImgUrl(String imgUrl, Long memberIdx) throws Exception {
-		AuthenticatedRequest authenticatedRequest = getPreAuth(imgUrl);
+	public String getPublicImgUrl(Long imageIdx, Long memberIdx) throws Exception {
+		ObjectStorage client = getClient();
+		Image img = imageRepository.findImageByIdx(imageIdx).orElseThrow(
+				() -> new ErrorHandler(ErrorStatus.IMAGE_NOT_FOUND)
+		);
+		AuthenticatedRequest authenticatedRequest = getPreAuth(img.getImgUrl());
+		
+		// 응답 로그
 		GetPreauthenticatedRequestRequest request =
 				GetPreauthenticatedRequestRequest.builder()
 						.namespaceName(BUCKET_NAME_SPACE)
@@ -90,31 +125,43 @@ public class OracleImageService implements ImageService {
 						.build();
 		GetPreauthenticatedRequestResponse response = client.getPreauthenticatedRequest(request);
 		log.info("response = " + response);
+
+		Member member = memberRepository.findByIdx(memberIdx).orElseThrow(() -> new ErrorHandler(ErrorStatus.MEMBER_NOT_FOUND));
+		addAccessUriToMember(member, authenticatedRequest, img.getImgUrl());
+
 		log.info("PublicImgUrl 발급에 성공하였습니다 : {}", DEFAULT_URI_PREFIX + authenticatedRequest.getAccessUri());
+		client.close();
 		return DEFAULT_URI_PREFIX +authenticatedRequest.getAccessUri();
 	}
 
 	@Override
-	public MultipartFile downloadImg(String imgUrl, Long memberIdx) throws Exception{
+	public MultipartFile downloadImg(Long imageIdx, Long memberIdx) throws Exception{
 		return null;
 	}
 
-	// 버킷에서 이미지 삭제
+	// 버킷에서 이미지와 인증정보 삭제
 	@Override
-	public void deleteImg(String imgUrl) throws Exception {
+	public void deleteImg(Image image) throws Exception {
+		ObjectStorage client = getClient();
 		DeleteObjectRequest request =
 				DeleteObjectRequest.builder()
 						.bucketName(BUCKET_NAME)
 						.namespaceName(BUCKET_NAME_SPACE)
-						.objectName(imgUrl)
+						.objectName(image.getImgUrl())
 						.build();
 
+		deletePreAuth(image.getParId());
+		imageRepository.delete(image);
+
 		client.deleteObject(request);
-//		client.close();
+		client.close();
 	}
 
 	// 오라클 버킷으로 파일 업로드
-	public Long upload(File uploadFile, String dirName, Member member) throws Exception{
+	public Long upload(File uploadFile, String dirName, Member member) throws Exception {
+		ObjectStorage client = getClient();
+		UploadManager uploadManager = getManager(client);
+
 		String fileName = dirName + UUID.randomUUID() + uploadFile.getName();   // S3에 저장된 파일 이름
 		String contentType = "img/" + fileName.substring(fileName.length() - 3); // PNG, JPG 만 가능함
 		String contentEncoding = null;
@@ -136,27 +183,42 @@ public class OracleImageService implements ImageService {
 		UploadResponse response = uploadManager.upload(uploadDetails);
 		log.info("Upload Success. File : {}", fileName);
 
+		client.close();
 		removeNewFile(uploadFile);
-//		client.close();
 		return saveImageToMember(member, fileName);
 	}
 
 	private Long saveImageToMember(Member member, String fileName) {
 		Image image = Image.builder()
 				.member(member)
-				.accessUri(fileName)
+				.imgUrl(fileName)
+				.imageType(ImageType.PROFILE)
 				.build();
+		if (fileName.contains(KAKAO_IMG_DIR)) {
+			image.updateImageType(ImageType.KAKAO);
+		}
 		imageRepository.save(image);
 		member.getImages().add(image);
 		return image.getIdx();
 	}
 
+	private void addAccessUriToMember(Member member, AuthenticatedRequest request, String imgUrl) {
+		List<Image> images = member.getImages();
+		for (Image img : images) {
+			if (img.getImgUrl().equals(imgUrl)) {
+				img.updateAccessUri(request.getAccessUri());
+				img.updateParId(request.authenticateId);
+			}
+		}
+	}
+
 
 	// 로컬에 파일 업로드 해서 convert
 	private Optional<File> convert(MultipartFile file) throws IOException {
-		File convertFile = new File(System.getProperty("user.home") + "/" + file.getOriginalFilename());
+		File convertFile = new File(System.getProperty("user.home") + "/rideTogetherDummy/" + UUID.randomUUID() +file.getOriginalFilename());
 		if (convertFile.createNewFile()) { // 바로 위에서 지정한 경로에 File이 생성됨 (경로가 잘못되었다면 생성 불가능)
-			try (FileOutputStream fos = new FileOutputStream(convertFile)) { // FileOutputStream 데이터를 파일에 바이트 스트림으로 저장하기 위함
+			try (FileOutputStream fos = new FileOutputStream(
+					convertFile)) { // FileOutputStream 데이터를 파일에 바이트 스트림으로 저장하기 위함
 				fos.write(file.getBytes());
 			}
 			return Optional.of(convertFile);
@@ -165,20 +227,28 @@ public class OracleImageService implements ImageService {
 	}
 
 	// 로컬에 저장된 이미지 지우기
-	private void removeNewFile(File targetFile) {
-		if (targetFile.delete()) {
-			log.info("File delete success");
-			return;
+	public void removeNewFile(File targetFile) {
+		log.info("@@@@@@@@ 지울 대상 파일 이름"+targetFile.getName());
+		log.info("@@@@@@@@ 지울 대상 파일 경로"+targetFile.getPath());
+		if (targetFile.exists()) {
+			if (targetFile.delete()) {
+				log.info("@@@@@@@@ File delete success");
+				return;
+			}
+			log.info("@@@@@@@@ File delete fail.");
 		}
-		log.info("File delete fail");
+		log.info("@@@@@@@@ File not exist.");
 	}
 
 	public AuthenticatedRequest getPreAuth(String imgUrl) throws Exception{
+		ObjectStorage client = getClient();
+
 		Calendar cal = Calendar.getInstance();
 		cal.set(2024, Calendar.DECEMBER, 30);
 
 		Date expireTime = cal.getTime();
 
+		log.info("권한을 얻어오기 위해 시도중입니다. 파일 이름 : {}", imgUrl);
 		CreatePreauthenticatedRequestDetails details =
 				CreatePreauthenticatedRequestDetails.builder()
 						.accessType(AccessType.ObjectReadWrite)
@@ -195,14 +265,15 @@ public class OracleImageService implements ImageService {
 						.build();
 
 		CreatePreauthenticatedRequestResponse response = client.createPreauthenticatedRequest(request);
-
+		client.close();
 		return AuthenticatedRequest.builder()
 				.authenticateId(response.getPreauthenticatedRequest().getId())
 				.accessUri(response.getPreauthenticatedRequest().getAccessUri())
 				.build();
 	}
 
-	public void deletePreAuth(String parId) throws Exception {
+	private void deletePreAuth(String parId) throws Exception {
+		ObjectStorage client = getClient();
 		DeletePreauthenticatedRequestRequest request =
 				DeletePreauthenticatedRequestRequest.builder()
 						.namespaceName(BUCKET_NAME_SPACE)
@@ -211,6 +282,7 @@ public class OracleImageService implements ImageService {
 						.build();
 
 		client.deletePreauthenticatedRequest(request);
+		client.close();
 	}
 
 	@Data
